requirements.txt
backcall==0.1.0
certifi==2019.9.11
cffi==1.13.2
cycler==0.10.0
decorator==4.4.1
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.3
olefile==0.46
pandas==0.25.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.0
six==1.13.0
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
